requirements.txt
transformers==4.30.2
torch==2.0.1
torchvision==0.15.2
tensorboard==2.13.0
Pillow==9.5.0
datasets==2.14.4
tqdm==4.65.0
numpy==1.25.2
pandas==2.1.4
scikit-learn==1.3.1
nltk==3.8.1
jieba==0.42.1
rouge==1.0.1
glob2==0.7.0
random
json
torchvision